Improve parsing color
- Use `tryParseColor` or `parseColorWithCache` where necessary.
