+++ src/inject/dynamic-theme/modify-css.ts
@@ -1,5 +1,4 @@
-import type {RGBA} from '../../utils/color';
-import {parse, rgbToHSL, hslToString} from '../../utils/color';
+import {rgbToHSL, hslToString, colorParseCache, tryParseColor, parseColorWithCache} from '../../utils/color';
 import {clamp} from '../../utils/math';
 import {getMatches} from '../../utils/text';
 import {getAbsoluteURL} from '../../utils/url';
@@ -100,7 +99,7 @@ export function getSelectionColor(theme: Theme) {
         backgroundColorSelection = modifyBackgroundColor({r: 0, g: 96, b: 212}, {...theme, grayscale: 0});
         foregroundColorSelection = modifyForegroundColor({r: 255, g: 255, b: 255}, {...theme, grayscale: 0});
     } else {
-        const rgb = parse(theme.selectionColor);
+        const rgb = parseColorWithCache(theme.selectionColor);
         const hsl = rgbToHSL(rgb);
         backgroundColorSelection = theme.selectionColor;
         if (hsl.l < 0.5) {
@@ -142,7 +141,7 @@ function getModifiedScrollbarStyle(theme: Theme) {
         colorThumbActive = modifyBackgroundColor({r: 96, g: 96, b: 96}, theme);
         colorCorner = modifyBackgroundColor({r: 255, g: 255, b: 255}, theme);
     } else {
-        const rgb = parse(theme.scrollbarColor);
+        const rgb = parseColorWithCache(theme.scrollbarColor);
         const hsl = rgbToHSL(rgb);
         const isLight = hsl.l > 0.5;
         const lighten = (lighter: number) => ({...hsl, l: clamp(hsl.l + lighter, 0, 1)});
@@ -196,32 +195,12 @@ const unparsableColors = new Set([
     'unset',
 ]);
 
-const colorParseCache = new Map<string, RGBA>();
-
-export function parseColorWithCache($color: string) {
-    $color = $color.trim();
-    if (colorParseCache.has($color)) {
-        return colorParseCache.get($color);
-    }
-    const color = parse($color);
-    colorParseCache.set($color, color);
-    return color;
-}
-
-function tryParseColor($color: string) {
-    try {
-        return parseColorWithCache($color);
-    } catch (err) {
-        return null;
-    }
-}
-
 function getColorModifier(prop: string, value: string): string | CSSValueModifier {
     if (unparsableColors.has(value.toLowerCase())) {
         return value;
     }
-    try {
-        const rgb = parseColorWithCache(value);
+    const rgb = tryParseColor(value);
+    if (rgb) {
         if (prop.indexOf('background') >= 0) {
             return (filter) => modifyBackgroundColor(rgb, filter);
         }
@@ -230,8 +209,8 @@ function getColorModifier(prop: string, value: string): string | CSSValueModifie
         }
         return (filter) => modifyForegroundColor(rgb, filter);
 
-    } catch (err) {
-        logWarn('Color parse error', err);
+    } else {
+        logWarn('Color parse error', value);
         return null;
     }
 }
